requirements.txt
Flask
Flask_Cors
numpy
openai
Pillow
python-dotenv
tensorflow
